Implement skeleton for a basic drawing framework
This is supposed to replace ratatui with an API of my own over the
crossterm library. By having it written myself I can better handle
any kind of synchronization issue when it comes to the terminal.

TODO: Generation of pixels to draw, done by reordering by line, then by
column and then by inserting newlines between the last pixel of a line
and the first of the next line.